works without buffered reading of FD; remove comparison between signed and unsigned variables

+++ src/spkdb_http.h
@@ -52,7 +52,7 @@ int create_parsed_http_req(parsed_http_req_t *parsed_http_req)
 
 void http_handle_res(fd_buff_struct_t *fd_conn)
 {
-	fprintf(stdout, "Sending response of %d bytes to FD %d\n", fd_conn->wbuff_size, fd_conn->fd);
+	fprintf(stdout, "Sending response of %ld bytes to FD %d\n", fd_conn->wbuff_size, fd_conn->fd);
 	ssize_t rv = write(fd_conn->fd, fd_conn->wbuff, fd_conn->wbuff_size);
 	fd_conn->state = STATE_END;
 }
@@ -131,7 +131,7 @@ void http_write_resp(char *resp, char *status, char *resp_headers, char *resp_bo
 	}
 
 	/* Get number of digits */
-	int num_digits = 1;
+	size_t num_digits = 1;
 	for (size_t i = 31; i > 0; --i) {
 		if (str_content_length[i] != '0') {
 			num_digits = i + 1;
